KSENROLL-908: fill out Assignment Info section in the inquiry page.
git-svn-id: https://svn.kuali.org/repos/student/branches/ks-1.3@29963 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-enroll/ks-enroll-ui/src/main/java/org/kuali/student/enrollment/class2/appointment/dto/AppointmentWindowWrapper.java
@@ -18,7 +18,7 @@ public class AppointmentWindowWrapper extends TimeSetWrapper {
     //Assignment Info
     Integer numberOfStudents;
     Integer numberOfSlots;
-    Integer meanStudentsPerSlot;
+    Float meanStudentsPerSlot;
     Date lastSlotPopulated;
     Date assignmentsCreated;
 
@@ -114,11 +114,11 @@ public class AppointmentWindowWrapper extends TimeSetWrapper {
         this.numberOfSlots = numberOfSlots;
     }
 
-    public Integer getMeanStudentsPerSlot() {
+    public Float getMeanStudentsPerSlot() {
         return meanStudentsPerSlot;
     }
 
-    public void setMeanStudentsPerSlot(Integer meanStudentsPerSlot) {
+    public void setMeanStudentsPerSlot(Float meanStudentsPerSlot) {
         this.meanStudentsPerSlot = meanStudentsPerSlot;
     }
